+++ common/src/main/java/me/rufia/fightorflight/net/CobblemonFightOrFlightNetwork.java
@@ -1,13 +1,17 @@
 package me.rufia.fightorflight.net;
 
 import dev.architectury.networking.NetworkManager;
+import me.rufia.fightorflight.net.handler.FOFStartBattlePokemonHandler;
+import me.rufia.fightorflight.net.handler.FOFStartBattleRequestHandler;
 import me.rufia.fightorflight.net.handler.SendCommandHandler;
 import me.rufia.fightorflight.net.handler.SendMoveSlotHandler;
+import me.rufia.fightorflight.net.packet.FOFStartBattlePokemonPacket;
+import me.rufia.fightorflight.net.packet.FOFStartBattleRequestPacket;
 import me.rufia.fightorflight.net.packet.SendCommandPacket;
 import me.rufia.fightorflight.net.packet.SendMoveSlotPacket;
 
 public class CobblemonFightOrFlightNetwork {
-    public static void init(){
+    public static void init() {
         NetworkManager.registerReceiver(NetworkManager.Side.C2S, SendCommandPacket.TYPE, SendCommandPacket.STREAM_CODEC, ((packet, context) -> {
             SendCommandHandler handler = new SendCommandHandler();
             handler.handle(packet, context);
@@ -16,5 +20,13 @@ public class CobblemonFightOrFlightNetwork {
             SendMoveSlotHandler handler = new SendMoveSlotHandler();
             handler.handle(packet, context);
         }));
+        NetworkManager.registerReceiver(NetworkManager.Side.S2C, FOFStartBattleRequestPacket.TYPE, FOFStartBattleRequestPacket.STREAM_CODEC, (packet, context) -> {
+            FOFStartBattleRequestHandler handler = new FOFStartBattleRequestHandler();
+            handler.handle(packet, context);
+        });
+        NetworkManager.registerReceiver(NetworkManager.Side.C2S, FOFStartBattlePokemonPacket.TYPE, FOFStartBattlePokemonPacket.STREAM_CODEC, (packet, context) -> {
+            FOFStartBattlePokemonHandler handler = new FOFStartBattlePokemonHandler();
+            handler.handle(packet, context);
+        });
     }
 }